design test app context and pkgutil to parse a package

+++ common/src/main/java/com/microsoft/hydralab/common/util/PkgUtil.java
@@ -8,33 +8,45 @@ import com.dd.plist.NSDictionary;
 import com.dd.plist.NSString;
 import com.dd.plist.PropertyListParser;
 import com.microsoft.hydralab.common.entity.common.AgentUpdateTask.TaskConst;
+import com.microsoft.hydralab.common.entity.common.AppComponent;
 import com.microsoft.hydralab.common.entity.common.EntityType;
 import com.microsoft.hydralab.common.entity.common.StorageFileInfo.ParserKey;
+import com.microsoft.hydralab.common.entity.common.TestAppContext;
 import net.dongliu.apk.parser.ApkFile;
 import net.dongliu.apk.parser.bean.ApkMeta;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.Assert;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
 
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.BufferUnderflowException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Enumeration;
 import java.util.List;
+import java.util.Objects;
 import java.util.Properties;
+import java.util.function.Consumer;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
 public class PkgUtil {
     static Logger classLogger = LoggerFactory.getLogger(PkgUtil.class);
+
     public static JSONObject analysisFile(File file, EntityType entityType) {
         JSONObject res = new JSONObject();
         switch (entityType) {
@@ -57,6 +69,118 @@ public class PkgUtil {
         return res;
     }
 
+    public static String getAndroidPackageManifest(File apkFile) throws IOException {
+        try (ApkFile apkFileObj = new ApkFile(apkFile)) {
+            return apkFileObj.getManifestXml();
+        }
+    }
+
+    public static List<String> getAndroidPackageComponents(File apkFile) {
+        return getAndroidPackageManifestTagAttrVal(apkFile, "android:name", "activity", "service", "receiver", "provider");
+    }
+
+    public static List<String> getAndroidPackageActivities(File apkFile) {
+        return getAndroidPackageManifestTagAttrVal(apkFile, "android:name", "activity");
+    }
+
+    public static List<String> getAndroidPackageManifestTagAttrVal(File apkFile, String attrName, String... tags) {
+        try {
+            // get the manifest xml from the apk file and get all the activity nodes
+            String manifestXml = getAndroidPackageManifest(apkFile);
+            return getAndroidPackageManifestTagAttrVal(manifestXml, attrName, tags);
+        } catch (Exception e) {
+            classLogger.error(e.getMessage(), e);
+        }
+        return null;
+    }
+
+    public static List<String> getAndroidPackageManifestTagAttrVal(String manifestXml, String attrName, String... tags) {
+        List<String> values = null;
+        try (ByteArrayInputStream bais = new ByteArrayInputStream(manifestXml.getBytes(StandardCharsets.UTF_8))) {
+            // get the manifest xml from the apk file and get all the activity nodes
+            DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+            // 创建一个 DocumentBuilder 对象
+            DocumentBuilder builder = factory.newDocumentBuilder();
+            // 使用DocumentBuilder对象解析XML
+            Document document = builder.parse(bais);
+            // 标准化XML，可选步骤
+            document.getDocumentElement().normalize();
+            // 获取所有名为"activity"的节点
+            List<Element> elements = getElementListByTags(document, tags);
+            values = new ArrayList<>();
+            // 遍历每一个节点
+            for (Element element : elements) {
+                String activityName = element.getAttribute(attrName);
+                values.add(activityName);
+            }
+        } catch (Exception e) {
+            classLogger.error(e.getMessage(), e);
+        }
+        return values;
+    }
+
+    public static void handleAndroidPackageManifestElementsByTags(File apkFile, Consumer<Element> consumer, String... tags) {
+        try {
+            // get the manifest xml from the apk file and get all the activity nodes
+            String manifestXml = getAndroidPackageManifest(apkFile);
+            try (ByteArrayInputStream bais = new ByteArrayInputStream(manifestXml.getBytes(StandardCharsets.UTF_8))) {
+                DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+                // 创建一个 DocumentBuilder 对象
+                DocumentBuilder builder = factory.newDocumentBuilder();
+                // 使用DocumentBuilder对象解析XML
+                Document document = builder.parse(bais);
+                // 标准化XML，可选步骤
+                document.getDocumentElement().normalize();
+                // 获取所有名为"activity"的节点
+                List<Element> elements = getElementListByTags(document, tags);
+                // 遍历每一个节点
+                for (Element element : elements) {
+                    consumer.accept(element);
+                }
+            }
+        } catch (Exception e) {
+            classLogger.error(e.getMessage(), e);
+        }
+    }
+
+    public static void handleAndroidPackageManifestElementsByTags(String manifestXml, Consumer<Element> consumer, String... tags) {
+        try (ByteArrayInputStream bais = new ByteArrayInputStream(manifestXml.getBytes(StandardCharsets.UTF_8))) {
+            DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+            // 创建一个 DocumentBuilder 对象
+            DocumentBuilder builder = factory.newDocumentBuilder();
+            // 使用DocumentBuilder对象解析XML
+            Document document = builder.parse(bais);
+            // 标准化XML，可选步骤
+            document.getDocumentElement().normalize();
+            // 获取所有名为"activity"的节点
+            List<Element> elements = getElementListByTags(document, tags);
+            // 遍历每一个节点
+            for (Element element : elements) {
+                consumer.accept(element);
+            }
+        } catch (Exception e) {
+            classLogger.error(e.getMessage(), e);
+        }
+    }
+
+    private static List<Element> getElementListByTags(Document document, String... tags) {
+        Assert.notNull(tags, "Element tags can not be null");
+        ArrayList<Element> elements = new ArrayList<>();
+        for (String tag : tags) {
+            if (tag == null) {
+                continue;
+            }
+            NodeList nodeList = document.getElementsByTagName(tag);
+            for (int i = 0; i < nodeList.getLength(); i++) {
+                Node node = nodeList.item(i);
+                if (node instanceof Element) {
+                    elements.add((Element) node);
+                }
+            }
+        }
+        return elements;
+    }
+
     private static JSONObject getAgentVersionFromJarFile(File file) {
         JSONObject res = new JSONObject();
         InputStream propertyStream = null;
@@ -72,16 +196,16 @@ public class PkgUtil {
             res.put(TaskConst.PARAM_VERSION_NAME, prop.getProperty(TaskConst.PROPERTY_VERSION_NAME));
             res.put(TaskConst.PARAM_VERSION_CODE, prop.getProperty(TaskConst.PROPERTY_VERSION_CODE));
         } catch (Exception e) {
-            e.printStackTrace();
+            classLogger.error(e.getMessage(), e);
         } finally {
             if (propertyStream != null) {
                 try {
                     propertyStream.close();
                 } catch (IOException e) {
-                    e.printStackTrace();
+                    classLogger.error(e.getMessage(), e);
                 }
             }
-            if (zipFile.exists()) {
+            if (zipFile != null && zipFile.exists()) {
                 zipFile.delete();
             }
         }
@@ -97,12 +221,8 @@ public class PkgUtil {
             res.put(ParserKey.VERSION, apkMeta.getVersionName());
             res.put(ParserKey.MIN_SDK_VERSION, apkMeta.getMinSdkVersion());
             res.put(ParserKey.TARGET_SDK_VERSION, apkMeta.getTargetSdkVersion());
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (BufferUnderflowException e) {
-            e.printStackTrace();
         } catch (Exception e) {
-            e.printStackTrace();
+            classLogger.error(e.getMessage(), e);
         }
         return res;
     }
@@ -110,7 +230,6 @@ public class PkgUtil {
     private static JSONObject analysisIpaFile(File ipa) {
         JSONObject res = new JSONObject();
         try {
-            String name, pkgName, version;
             File zipFile = convertToZipFile(ipa, FILE_SUFFIX.IPA_FILE);
             Assert.notNull(zipFile, "Convert .ipa file to .zip file failed.");
             File file = getPlistFromZip(zipFile, zipFile.getParent());
@@ -118,7 +237,7 @@ public class PkgUtil {
             Assert.notNull(file, "Analysis .ipa file failed.");
             analysisPlist(file, res);
         } catch (Exception e) {
-            e.printStackTrace();
+            classLogger.error(e.getMessage(), e);
         }
         return res;
     }
@@ -141,12 +260,12 @@ public class PkgUtil {
             } else if (pyMainFile != null) {
                 res.put(ParserKey.APP_NAME, "Python Runner");
                 res.put(ParserKey.PKG_NAME, "Python Runner");
-            } else if (yamlFiles.size() == 0) {
+            } else if (yamlFiles.isEmpty()) {
                 classLogger.warn("Analysis .zip file failed. It's not a valid XCTEST package, Maestro case or Python package.");
             }
             FileUtil.deleteFile(unzippedFolder);
         } catch (Exception e) {
-            e.printStackTrace();
+            classLogger.error(e.getMessage(), e);
         }
         return res;
 
@@ -189,7 +308,7 @@ public class PkgUtil {
         InputStream input = null;
         OutputStream output = null;
         File result = null;
-        File unzipFile = null;
+        File unzipFile;
         ZipFile zipFile = null;
         try {
             //Create zip file object
@@ -204,15 +323,15 @@ public class PkgUtil {
             //Get zip file entry enumeration object
             Enumeration<? extends ZipEntry> zipEnum = zipFile.entries();
             //define object
-            ZipEntry entry = null;
-            String entryName = null;
-            String[] names = null;
+            ZipEntry entry;
+            String entryName;
+            String[] names;
             int length;
             //loop reading entries
             while (zipEnum.hasMoreElements()) {
                 //get the current entry
                 entry = zipEnum.nextElement();
-                entryName = new String(entry.getName());
+                entryName = entry.getName();
                 //separate entry names with/
                 names = entryName.split("\\/");
                 length = names.length;
@@ -222,7 +341,7 @@ public class PkgUtil {
                         result = new File(unzipFile.getAbsolutePath() + "/Info.plist");
                         output = Files.newOutputStream(result.toPath());
                         byte[] buffer = new byte[1024 * 8];
-                        int readLen = 0;
+                        int readLen;
                         while ((readLen = input.read(buffer, 0, 1024 * 8)) != -1) {
                             output.write(buffer, 0, readLen);
                         }
@@ -230,8 +349,8 @@ public class PkgUtil {
                     }
                 }
             }
-        } catch (Exception ex) {
-            ex.printStackTrace();
+        } catch (Exception e) {
+            classLogger.error(e.getMessage(), e);
         } finally {
             if (input != null) {
                 input.close();
@@ -278,7 +397,7 @@ public class PkgUtil {
 
     private static File convertToZipFile(File file, String suffix) {
         try {
-            int bytes = 0;
+            int bytes;
             String filename = file.getAbsolutePath().replaceAll(suffix, FILE_SUFFIX.ZIP_FILE);
             File zipFile = new File(filename);
             if (file.exists()) {
@@ -295,9 +414,33 @@ public class PkgUtil {
                 return zipFile;
             }
         } catch (Exception e) {
-            e.printStackTrace();
+            classLogger.error(e.getMessage(), e);
         }
         return null;
+    }
+
+    public static TestAppContext getTestAppContext(File apkFile) throws IOException {
+        try (ApkFile apkFileObj = new ApkFile(apkFile)) {
+            TestAppContext testAppContext = new TestAppContext();
+            ApkMeta apkMeta = apkFileObj.getApkMeta();
+            testAppContext.setAppName(apkMeta.getName());
+            testAppContext.setPackageName(apkMeta.getPackageName());
+            testAppContext.setPackageSize(apkFile.length());
+            testAppContext.setAppVersion(apkMeta.getVersionName());
+            handleAndroidPackageManifestElementsByTags(apkFileObj.getManifestXml(), element -> {
+                String tagName = element.getTagName();
+                String attributeNameVal = element.getAttribute("android:name");
+                if (Objects.equals(tagName, "activity")) {
+                    testAppContext.getAppComponents().add(new AppComponent(attributeNameVal));
+                } else if (Objects.equals(tagName, "provider")) {
+                    testAppContext.getAppComponents().add(new AppComponent(attributeNameVal, AppComponent.Type.DATASOURCE));
+                } else {
+                    testAppContext.getAppComponents().add(new AppComponent(attributeNameVal, AppComponent.Type.SERVICE));
+                }
+            }, "activity", "service", "receiver", "provider");
+
+            return testAppContext;
+        }
     }
 
     public interface FILE_SUFFIX {
@@ -307,6 +450,6 @@ public class PkgUtil {
         String ZIP_FILE = ".zip";
         String IPA_FILE = ".ipa";
         String JSON_FILE = ".json";
-        String APP_FILE = ".app";
+//        String APP_FILE = ".app";
     }
 }
